Add files via upload
This is a MATLAB implementation of the non-negative least squares algorithm. While MATLAB has its own implementation of non-negative least squares called lsqnonneg.  I decided to try implementing it on my own for educational purposes. I hope this will help  others understand how the algorithm works.
The algorithm is described in many papers and textbooks. The original description of the algorithm is found in the book "Solving Least Squares Problems" by Lawson and Hanson. 

Non-negative least squares is an active set technique. The idea of the active set technique, is that, typically only a few coefficients are actively violating the non-negativity constraint. It iteratively identifies the coefficients that are NOT violating the constraints and moves them from the active set to passive set. It then estimates the non-negative signal using unconstrained least squares estimation using only the coefficients from the passive set.